Rev. 3: Update functionality to get image count method

+++ ex2_hal/iris/hardware_interface/include/iris.h
@@ -88,7 +88,7 @@ typedef struct __attribute__((__packed__)) {
 IrisHALReturn iris_take_pic();
 IrisHALReturn iris_get_image_length(uint32_t *image_length);
 IrisHALReturn iris_transfer_image();
-IrisHALReturn iris_get_image_count(uint8_t *image_count);
+IrisHALReturn iris_get_image_count(uint16_t *image_count);
 IrisHALReturn iris_toggle_sensor_idle(uint8_t toggle);
 IrisHALReturn iris_get_housekeeping(iris_housekeeping_data hk_data);
 IrisHALReturn iris_update_sensor_i2c_reg();
